Moved 'gui' files to 'src' directory, renamed to 'drive_sync'
